requirements.txt
asgiref==3.4.1
autopep8==1.5.7
Babel==2.9.1
backports.entry-points-selectable==1.1.0
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
click==8.0.1
defusedxml==0.7.1
distlib==0.3.3
dj-database-url==0.5.0
Django==2.2.25
django-allauth==0.41.0
django-babel==0.6.2
django-crispy-forms==1.13.0
django-livereload-server==0.4
django-postgres-extensions==0.9.3
djangorestframework==3.12.4
dnspython==2.1.0
filelock==3.3.0
Flask==2.0.1
Flask-PyMongo==2.3.0
google-auth==1.34.0
google-auth-oauthlib==0.4.5
grunt-django==0.2.4
gspread==4.0.0
gunicorn==20.1.0
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
jsonfield==3.1.0
MarkupSafe==2.0.1
oauthlib==3.1.1
platformdirs==2.4.0
psycopg2-binary==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.7.0
pygame==2.0.1
pymongo==3.12.0
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
simple-term-menu==1.4.1
six==1.15.0
soupsieve==2.3.1
sqlparse==0.4.2
toml==0.10.2
tornado==6.1
urllib3==1.26.6
virtualenv==20.8.1
Werkzeug==2.0.1
